Seperated Tokenizers into different classes
- Addressed some comments discussed with Mircea.
