+++ zeeguu/core/tokenization/__init__.py
@@ -1,5 +1,8 @@
 from .token import Token
-from .tokenizer import ZeeguuTokenizer, TokenizerModel
+from .stanza_tokenizer import StanzaTokenizer
+from .nltk_tokenizer import NLTKTokenizer
+from .zeeguu_tokenizer import TokenizerModel
+
 
 """
 - NLTK is the fastest model.
@@ -7,3 +10,10 @@ from .tokenizer import ZeeguuTokenizer, TokenizerModel
 - STANZA_TOKEN_POS uses the same a TOKEN, but also does POS, much slower.
 """
 TOKENIZER_MODEL = TokenizerModel.STANZA_TOKEN_ONLY
+
+
+def get_tokenizer(language, model):
+    if model in StanzaTokenizer.STANZA_MODELS:
+        return StanzaTokenizer(language, model)
+    else:
+        return NLTKTokenizer(language)
